Divides provider code in multiple files, easy to navigate
everything in same file, was getting difficult to navigate
now, separates out detect and parsepayload code in separate files.
