+++ pkg/provider/gitea/detect_test.go
