+++ pkg/provider/gitlab/gitlab.go
@@ -3,7 +3,6 @@ package gitlab
 import (
 	"context"
 	"crypto/subtle"
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"path/filepath"
@@ -67,55 +66,6 @@ func (v *Provider) Validate(_ context.Context, _ *params.Run, event *info.Event)
 	return nil
 }
 
-// Detect processes event and detect if it is a gitlab event, whether to process or reject it
-// returns (if is a GL event, whether to process or reject, logger with event metadata,, error if any occurred)
-func (v *Provider) Detect(req *http.Request, payload string, logger *zap.SugaredLogger) (bool, bool, *zap.SugaredLogger, string, error) {
-	isGL := false
-	event := req.Header.Get("X-Gitlab-Event")
-	if event == "" {
-		return false, false, logger, "no gitlab event", nil
-	}
-
-	// it is a Gitlab event
-	isGL = true
-
-	setLoggerAndProceed := func(processEvent bool, reason string, err error) (bool, bool, *zap.SugaredLogger,
-		string, error,
-	) {
-		logger = logger.With("provider", "gitlab", "event-id", req.Header.Get("X-Request-Id"))
-		return isGL, processEvent, logger, reason, err
-	}
-
-	eventInt, err := gitlab.ParseWebhook(gitlab.EventType(event), []byte(payload))
-	if err != nil {
-		return setLoggerAndProceed(false, "", err)
-	}
-	_ = json.Unmarshal([]byte(payload), &eventInt)
-
-	switch gitEvent := eventInt.(type) {
-	case *gitlab.MergeEvent:
-		if provider.Valid(gitEvent.ObjectAttributes.Action, []string{"open", "update", "reopen"}) {
-			return setLoggerAndProceed(true, "", nil)
-		}
-		return setLoggerAndProceed(false, fmt.Sprintf("not a merge event we care about: \"%s\"",
-			gitEvent.ObjectAttributes.Action), nil)
-	case *gitlab.PushEvent:
-		return setLoggerAndProceed(true, "", nil)
-	case *gitlab.MergeCommentEvent:
-		if gitEvent.MergeRequest.State == "opened" {
-			if provider.IsTestRetestComment(gitEvent.ObjectAttributes.Note) {
-				return setLoggerAndProceed(true, "", nil)
-			}
-			if provider.IsOkToTestComment(gitEvent.ObjectAttributes.Note) {
-				return setLoggerAndProceed(true, "", nil)
-			}
-		}
-		return setLoggerAndProceed(false, "not a gitops style merge comment event", nil)
-	default:
-		return setLoggerAndProceed(false, "", fmt.Errorf("gitlab: event \"%s\" is not supported", event))
-	}
-}
-
 // If I understood properly, you can have "personal" projects and groups
 // attached projects. But this doesn't seem to show in the API, so we
 // are just doing it the path_with_namespace to get the "org".
@@ -128,109 +78,6 @@ func getOrgRepo(pathWithNamespace string) (string, string) {
 	org := filepath.Dir(pathWithNamespace)
 	org = strings.ReplaceAll(org, "/", "-")
 	return org, filepath.Base(pathWithNamespace)
-}
-
-func (v *Provider) ParsePayload(_ context.Context, _ *params.Run, request *http.Request,
-	payload string,
-) (*info.Event, error) {
-	// TODO: parse request to figure out which event
-	var processedEvent *info.Event
-
-	event := request.Header.Get("X-Gitlab-Event")
-	if event == "" {
-		return nil, fmt.Errorf("failed to find event type in request header")
-	}
-
-	payloadB := []byte(payload)
-	eventInt, err := gitlab.ParseWebhook(gitlab.EventType(event), payloadB)
-	if err != nil {
-		return nil, err
-	}
-	_ = json.Unmarshal(payloadB, &eventInt)
-
-	switch gitEvent := eventInt.(type) {
-	case *gitlab.MergeEvent:
-		processedEvent = info.NewEvent()
-		// Organization:  event.GetRepo().GetOwner().GetLogin(),
-		processedEvent.Sender = gitEvent.User.Username
-		processedEvent.DefaultBranch = gitEvent.Project.DefaultBranch
-		processedEvent.URL = gitEvent.Project.WebURL
-		processedEvent.SHA = gitEvent.ObjectAttributes.LastCommit.ID
-		processedEvent.SHAURL = gitEvent.ObjectAttributes.LastCommit.URL
-		processedEvent.SHATitle = gitEvent.ObjectAttributes.Title
-		processedEvent.HeadBranch = gitEvent.ObjectAttributes.SourceBranch
-		processedEvent.BaseBranch = gitEvent.ObjectAttributes.TargetBranch
-		processedEvent.PullRequestNumber = gitEvent.ObjectAttributes.IID
-		processedEvent.PullRequestTitle = gitEvent.ObjectAttributes.Title
-		v.targetProjectID = gitEvent.Project.ID
-		v.sourceProjectID = gitEvent.ObjectAttributes.SourceProjectID
-		v.userID = gitEvent.User.ID
-
-		v.pathWithNamespace = gitEvent.ObjectAttributes.Target.PathWithNamespace
-		processedEvent.Organization, processedEvent.Repository = getOrgRepo(v.pathWithNamespace)
-		processedEvent.TriggerTarget = "pull_request"
-		processedEvent.SourceProjectID = gitEvent.ObjectAttributes.SourceProjectID
-		processedEvent.TargetProjectID = gitEvent.Project.ID
-	case *gitlab.PushEvent:
-		if len(gitEvent.Commits) == 0 {
-			return nil, fmt.Errorf("no commits attached to this push event")
-		}
-		processedEvent = info.NewEvent()
-		processedEvent.Sender = gitEvent.UserUsername
-		processedEvent.DefaultBranch = gitEvent.Project.DefaultBranch
-		processedEvent.URL = gitEvent.Project.WebURL
-		processedEvent.SHA = gitEvent.Commits[0].ID
-		processedEvent.SHAURL = gitEvent.Commits[0].URL
-		processedEvent.SHATitle = gitEvent.Commits[0].Title
-		processedEvent.HeadBranch = gitEvent.Ref
-		processedEvent.BaseBranch = gitEvent.Ref
-		processedEvent.TriggerTarget = "push"
-		v.pathWithNamespace = gitEvent.Project.PathWithNamespace
-		processedEvent.Organization, processedEvent.Repository = getOrgRepo(v.pathWithNamespace)
-		v.targetProjectID = gitEvent.ProjectID
-		v.sourceProjectID = gitEvent.ProjectID
-		v.userID = gitEvent.UserID
-		processedEvent.SourceProjectID = gitEvent.ProjectID
-		processedEvent.TargetProjectID = gitEvent.ProjectID
-	case *gitlab.MergeCommentEvent:
-		processedEvent = info.NewEvent()
-		processedEvent.Sender = gitEvent.User.Username
-		processedEvent.DefaultBranch = gitEvent.Project.DefaultBranch
-		processedEvent.URL = gitEvent.Project.WebURL
-		processedEvent.SHA = gitEvent.MergeRequest.LastCommit.ID
-		processedEvent.SHAURL = gitEvent.MergeRequest.LastCommit.URL
-		// TODO: change this back to Title when we get this pr available merged https://github.com/xanzy/go-gitlab/pull/1406/files
-		processedEvent.SHATitle = gitEvent.MergeRequest.LastCommit.Message
-		processedEvent.BaseBranch = gitEvent.MergeRequest.TargetBranch
-		processedEvent.HeadBranch = gitEvent.MergeRequest.SourceBranch
-		// if it is a /test or /retest comment with pipelinerun name figure out the pipelineRun name
-		if provider.IsTestRetestComment(gitEvent.ObjectAttributes.Note) {
-			processedEvent.TargetTestPipelineRun = provider.GetPipelineRunFromComment(gitEvent.ObjectAttributes.Note)
-		}
-
-		v.pathWithNamespace = gitEvent.Project.PathWithNamespace
-		processedEvent.Organization, processedEvent.Repository = getOrgRepo(v.pathWithNamespace)
-		processedEvent.TriggerTarget = "pull_request"
-
-		processedEvent.PullRequestNumber = gitEvent.MergeRequest.IID
-		v.targetProjectID = gitEvent.MergeRequest.TargetProjectID
-		v.sourceProjectID = gitEvent.MergeRequest.SourceProjectID
-		v.userID = gitEvent.User.ID
-		processedEvent.SourceProjectID = gitEvent.MergeRequest.SourceProjectID
-		processedEvent.TargetProjectID = gitEvent.MergeRequest.TargetProjectID
-	default:
-		return nil, fmt.Errorf("event %s is not supported", event)
-	}
-
-	processedEvent.Event = eventInt
-
-	// Remove the " Hook" suffix so looks better in status, and since we don't
-	// really use it anymore we good to do whatever we want with it for
-	// cosmetics.
-	processedEvent.EventType = strings.ReplaceAll(event, " Hook", "")
-
-	v.repoURL = processedEvent.URL
-	return processedEvent, nil
 }
 
 func (v *Provider) GetConfig() *info.ProviderConfig {
