+++ pkg/provider/gitlab/gitlab_test.go
@@ -9,7 +9,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/openshift-pipelines/pipelines-as-code/pkg/params"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/params/info"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/params/settings"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/provider"
@@ -177,140 +176,6 @@ func TestCreateStatus(t *testing.T) {
 	}
 }
 
-func TestParsePayload(t *testing.T) {
-	sample := thelp.TEvent{
-		Username:          "foo",
-		DefaultBranch:     "main",
-		URL:               "https://foo.com",
-		SHA:               "sha",
-		SHAurl:            "https://url",
-		SHAtitle:          "commit it",
-		Headbranch:        "branch",
-		Basebranch:        "main",
-		UserID:            10,
-		MRID:              1,
-		TargetProjectID:   100,
-		SourceProjectID:   200,
-		PathWithNameSpace: "hello/this/is/me/ze/project",
-	}
-	type fields struct {
-		targetProjectID int
-		sourceProjectID int
-		userID          int
-	}
-	type args struct {
-		event   gitlab.EventType
-		payload string
-	}
-	tests := []struct {
-		name       string
-		fields     fields
-		args       args
-		want       *info.Event
-		wantErr    bool
-		wantClient bool
-	}{
-		{
-			name: "bad payload",
-			args: args{
-				payload: "nono",
-				event:   "none",
-			},
-			wantErr: true,
-		},
-		{
-			name: "event not supported",
-			args: args{
-				event:   gitlab.EventTypePipeline,
-				payload: sample.MREventAsJSON(),
-			},
-			wantErr: true,
-		},
-		{
-			name: "merge event",
-			args: args{
-				event:   gitlab.EventTypeMergeRequest,
-				payload: sample.MREventAsJSON(),
-			},
-			want: &info.Event{
-				EventType:     "Merge Request",
-				TriggerTarget: "pull_request",
-				Organization:  "hello-this-is-me-ze",
-				Repository:    "project",
-			},
-		},
-		{
-			name: "push event no commits",
-			args: args{
-				event:   gitlab.EventTypePush,
-				payload: sample.PushEventAsJSON(false),
-			},
-			wantErr: true,
-		},
-		{
-			name: "push event",
-			args: args{
-				event:   gitlab.EventTypePush,
-				payload: sample.PushEventAsJSON(true),
-			},
-			want: &info.Event{
-				EventType:     "Push",
-				TriggerTarget: "push",
-				Organization:  "hello-this-is-me-ze",
-				Repository:    "project",
-			},
-		},
-
-		{
-			name: "note event",
-			args: args{
-				event:   gitlab.EventTypeNote,
-				payload: sample.NoteEventAsJSON(""),
-			},
-			want: &info.Event{
-				EventType:     "Note",
-				TriggerTarget: "pull_request",
-				Organization:  "hello-this-is-me-ze",
-				Repository:    "project",
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx, _ := rtesting.SetupFakeContext(t)
-			v := &Provider{
-				Token:           gitlab.String("tokeneuneu"),
-				targetProjectID: tt.fields.targetProjectID,
-				sourceProjectID: tt.fields.sourceProjectID,
-				userID:          tt.fields.userID,
-			}
-			if tt.wantClient {
-				client, _, tearDown := thelp.Setup(ctx, t)
-				v.Client = client
-				defer tearDown()
-			}
-			run := &params.Run{
-				Info: info.Info{},
-			}
-
-			request := &http.Request{Header: map[string][]string{}}
-			request.Header.Set("X-Gitlab-Event", string(tt.args.event))
-
-			got, err := v.ParsePayload(ctx, run, request, tt.args.payload)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ParsePayload() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if tt.want != nil {
-				assert.Equal(t, tt.want.TriggerTarget, got.TriggerTarget)
-				assert.Equal(t, tt.want.EventType, got.EventType)
-				assert.Equal(t, tt.want.Organization, got.Organization)
-				assert.Equal(t, tt.want.Repository, got.Repository)
-			}
-		})
-	}
-}
-
 func TestGetCommitInfo(t *testing.T) {
 	ctx, _ := rtesting.SetupFakeContext(t)
 	client, _, tearDown := thelp.Setup(ctx, t)
@@ -466,112 +331,6 @@ func TestGetFileInsideRepo(t *testing.T) {
 
 	_, err = v.GetFileInsideRepo(ctx, event, "notfound", "")
 	assert.Assert(t, err != nil)
-}
-
-func TestProvider_Detect(t *testing.T) {
-	sample := thelp.TEvent{
-		Username:          "foo",
-		DefaultBranch:     "main",
-		URL:               "https://foo.com",
-		SHA:               "sha",
-		SHAurl:            "https://url",
-		SHAtitle:          "commit it",
-		Headbranch:        "branch",
-		Basebranch:        "main",
-		UserID:            10,
-		MRID:              1,
-		TargetProjectID:   100,
-		SourceProjectID:   200,
-		PathWithNameSpace: "hello/this/is/me/ze/project",
-	}
-	tests := []struct {
-		name          string
-		wantErrString string
-		isGL          bool
-		processReq    bool
-		event         string
-		eventType     gitlab.EventType
-		wantReason    string
-	}{
-		{
-			name:       "not a gitlab Event",
-			eventType:  "",
-			isGL:       false,
-			processReq: false,
-		},
-		{
-			name:          "invalid gitlab Event",
-			eventType:     "invalid",
-			wantErrString: "unexpected event type: invalid",
-			isGL:          false,
-			processReq:    false,
-		},
-		{
-			name:       "valid merge Event",
-			event:      sample.MREventAsJSON(),
-			eventType:  gitlab.EventTypeMergeRequest,
-			isGL:       true,
-			processReq: true,
-		},
-		{
-			name:       "issue note event with no valid comment",
-			event:      sample.NoteEventAsJSON("abc"),
-			eventType:  gitlab.EventTypeNote,
-			isGL:       true,
-			processReq: false,
-		},
-		{
-			name:       "issue note Event with ok-to-test comment",
-			event:      sample.NoteEventAsJSON("/ok-to-test"),
-			eventType:  gitlab.EventTypeNote,
-			isGL:       true,
-			processReq: true,
-		},
-		{
-			name:       "issue comment Event with ok-to-test and some string",
-			event:      sample.NoteEventAsJSON("abc /ok-to-test"),
-			eventType:  gitlab.EventTypeNote,
-			isGL:       true,
-			processReq: false,
-		},
-		{
-			name:       "issue comment Event with retest",
-			event:      sample.NoteEventAsJSON("/retest"),
-			eventType:  gitlab.EventTypeNote,
-			isGL:       true,
-			processReq: true,
-		},
-		{
-			name:       "push event",
-			event:      sample.PushEventAsJSON(true),
-			eventType:  gitlab.EventTypePush,
-			isGL:       true,
-			processReq: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			gprovider := Provider{}
-			logger := getLogger()
-
-			header := http.Header{}
-			header.Set("X-Gitlab-Event", string(tt.eventType))
-			req := &http.Request{Header: header}
-			isGL, processReq, _, reason, err := gprovider.Detect(req, tt.event, logger)
-			if tt.wantErrString != "" {
-				assert.ErrorContains(t, err, tt.wantErrString)
-				return
-			}
-			assert.NilError(t, err)
-			if tt.wantReason != "" {
-				assert.Assert(t, strings.Contains(reason, tt.wantReason))
-				return
-			}
-			assert.Equal(t, tt.isGL, isGL)
-			assert.Equal(t, tt.processReq, processReq)
-		})
-	}
 }
 
 func getLogger() *zap.SugaredLogger {
